+++ src/main/java/org/accompany/backend/domain/survey/dto/request/SurveySaveReq.java
@@ -2,7 +2,7 @@ package org.accompany.backend.domain.survey.dto.request;
 
 import java.util.List;
 
-public record SurveyTempSaveReq (
+public record SurveySaveReq (
         List<SurveyAnswerIdReq> answers
 ){
 }
