+++ src/main/java/org/accompany/backend/domain/survey/service/SurveyService.java
@@ -2,7 +2,7 @@ package org.accompany.backend.domain.survey.service;
 
 import org.accompany.backend.domain.procedure.entity.Procedure;
 import org.accompany.backend.domain.survey.dto.response.SurveyListRes;
-import org.accompany.backend.domain.survey.dto.request.SurveyTempSaveReq;
+import org.accompany.backend.domain.survey.dto.request.SurveySaveReq;
 import org.accompany.backend.domain.survey.dto.response.SurveySubmitRes;
 import org.accompany.backend.domain.survey.dto.response.SurveyTempSaveRes;
 
@@ -12,8 +12,8 @@ import java.time.LocalDateTime;
 public interface SurveyService {
     SurveyListRes getSurveyList(Long userId);
     void skipSurvey(Long userId);
-    SurveyTempSaveRes saveTempSurvey(Long userId, SurveyTempSaveReq request);
-    SurveySubmitRes submitSurvey(Long userId, SurveyTempSaveReq request);
+    SurveyTempSaveRes saveTempSurvey(Long userId, SurveySaveReq request);
+    SurveySubmitRes submitSurvey(Long userId, SurveySaveReq request);
 
     LocalDateTime calculateDueDate(Procedure procedure, LocalDate dateOfDeath);
 }
